[Misc] Rename common_fused_moe to moe/fused_moe
