+++ vllm_ascend/ops/__init__.py
@@ -17,8 +17,8 @@
 
 import torch
 
-import vllm_ascend.ops.common_fused_moe  # noqa
 import vllm_ascend.ops.layernorm  # noqa
+import vllm_ascend.ops.moe.fused_moe  # noqa
 import vllm_ascend.ops.register_custom_ops  # noqa
 import vllm_ascend.ops.vocab_parallel_embedding  # noqa
 from vllm_ascend.ops.activation import AscendQuickGELU, AscendSiluAndMul
